Refactor code to add state of level to session
When a user starts the game a session is created
and assigned a level name. The found status of
Sonic, Tails and Knuckles exists on the session as
well as the time the session was created. This is
more intuitive to work with than having a seperate
table to keep track of level states and saves
having to track start time and end times for each
level referencing back to a single session.