Refactor createSession add pruning and shorter exp
Cookies and sessions now have a 1hr expiry as
opposed to a day. This should reduce the timeframe
in which the sessions are held in the database.
Adding pruning every 60 secs will help to keep the
sessions database clear of expired sessions.
createExpressSession has been refactored with the
option to enable regenerate. saveUninitialized has
been set to false as it will prevent sessions from
being overwritten in the case that the same sid
is generated - which it was. The old scenario
resulted in the sessions being overwritten even if
the user didn't have a cookie.